+++ source/Private/Get-CommentBasedHelp.ps1
@@ -1,27 +1,30 @@
 <#
     .SYNOPSIS
-        Get-ClassResourceCommentBasedHelp returns comment-based help for a DSC resource.
+        Get-CommentBasedHelp returns comment-based help from a PowerShell
+        script file.
 
     .DESCRIPTION
-        Get-ClassResourceCommentBasedHelp returns comment-based help for a DSC resource
-        by parsing the source PowerShell script file.
+        Get-CommentBasedHelp returns comment-based help for a PowerShell script
+        file by parsing the source PowerShell script file. If the comment block
+        is not the first element in the file then the first comment block will
+        be located and everything before it will be removed before parsing.
 
     .PARAMETER Path
-        The path to the DSC resource PowerShell script file. This should be a
-        PowerShell script file that only contain one resource with the
+        The path to the PowerShell script file. This should be a
+        PowerShell script file that contains one resource with the
         comment-based help at the top of the file.
 
     .OUTPUTS
         System.Management.Automation.Language.CommentHelpInfo
 
     .EXAMPLE
-        $commentBasedHelp = Get-ClassResourceCommentBasedHelp -Path 'c:\MyProject\source\Classes\010-MyResourceName.ps1'
+        $commentBasedHelp = Get-CommentBasedHelp -Path 'c:\MyProject\source\Classes\010-MyResourceName.ps1'
 
         This example parses the comment-based help from the PowerShell script file
         'c:\MyProject\source\Classes\010-MyResourceName.ps1' and returns an
         object of System.Management.Automation.Language.CommentHelpInfo.
 #>
-function Get-ClassResourceCommentBasedHelp
+function Get-CommentBasedHelp
 {
     [CmdletBinding()]
     [OutputType([System.Management.Automation.Language.CommentHelpInfo])]
@@ -42,11 +45,29 @@ function Get-ClassResourceCommentBasedHelp
     #>
     $Path = Resolve-Path -Path $Path
 
-    Write-Verbose -Message ($script:localizedData.ClassBasedCommentBasedHelpMessage -f $Path)
+    Write-Verbose -Message ($script:localizedData.CommentBasedHelpMessage -f $Path)
+
+    $scriptContent = Get-Content -Path $Path -Raw
+
+    # Ensure the comment-based help block is at the top of the file.
+    $regexOptions = [System.Text.RegularExpressions.RegexOptions]::Multiline
+    $firstCommentBlockStart = [System.Text.RegularExpressions.Regex]::Match($scriptContent, "^<#(\n|\r|\r\n)$", $regexOptions)
+
+    if ($firstCommentBlockStart.Success)
+    {
+        Write-Verbose -Message ($script:localizedData.CommentBasedHelpBlockNotAtTopMessage -f $Path)
+        $scriptContent = $scriptContent.Substring($firstCommentBlockStart.Index)
+    }
+    else
+    {
+        Write-Warning -Message (
+            $script:localizedData.CommentBasedHelpBlockNotFound -f $Path
+        )
+    }
 
     $tokens, $parseErrors = $null
 
-    $ast = [System.Management.Automation.Language.Parser]::ParseFile($Path, [ref] $tokens, [ref] $parseErrors)
+    $ast = [System.Management.Automation.Language.Parser]::ParseInput($scriptContent, [ref] $tokens, [ref] $parseErrors)
 
     if ($parseErrors)
     {