+++ CHANGELOG.md
@@ -27,6 +27,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Get-MofSchemaObject`
   - Refactored to reduce code duplication when adding functions for supporting
     composite resources.
+- `Get-ClassResourceCommentBasedHelp`
+  - Renamed this function to `Get-CommentBasedHelp` so that it made sense to
+    use with composite DSC resources.
+  - Enabled the function to extract the comment block if it is not at the top
+    of the script file to support composite resources.
 
 ### Fixed